split out usdx reward tests into own file
